fix: missAmount exceeds range of Integer resolved

+++ src/main/java/com/halggeol/backend/insight/dto/InsightDetailResponseDTO.java
@@ -35,7 +35,7 @@ public class InsightDetailResponseDTO {
 
     // 회고 관련
     private Integer regretScore; // 후회지수
-    private Integer missAmount; // 놓친 금액
+    private Long missAmount; // 놓친 금액
 
     // 상품 정보 관련
     // 현재는 비동기 처리로 사용하는 중
